Move motorConfig_s, motorDevConfig_s and associated PG handling into pg/motor.[ch]
