+++ src/main/blackbox/blackbox.c
@@ -45,6 +45,7 @@
 #include "config/feature.h"
 #include "pg/pg.h"
 #include "pg/pg_ids.h"
+#include "pg/motor.h"
 #include "pg/rx.h"
 
 #include "drivers/compass/compass.h"
