+++ src/main/cli/cli.c
@@ -130,6 +130,7 @@ uint8_t cliMode = 0;
 #include "pg/gyrodev.h"
 #include "pg/max7456.h"
 #include "pg/mco.h"
+#include "pg/motor.h"
 #include "pg/pinio.h"
 #include "pg/pg.h"
 #include "pg/pg_ids.h"
