+++ src/main/cms/cms_menu_misc.c
@@ -41,6 +41,7 @@
 #include "config/feature.h"
 #include "pg/pg.h"
 #include "pg/pg_ids.h"
+#include "pg/motor.h"
 #include "pg/rx.h"
 
 #include "fc/config.h"
