--- src/main/drivers/pwm_output_counts.h
@@ -1,34 +1,0 @@
-/*
- * This file is part of Cleanflight and Betaflight.
- *
- * Cleanflight and Betaflight are free software. You can redistribute
- * this software and/or modify this software under the terms of the
- * GNU General Public License as published by the Free Software
- * Foundation, either version 3 of the License, or (at your option)
- * any later version.
- *
- * Cleanflight and Betaflight are distributed in the hope that they
- * will be useful, but WITHOUT ANY WARRANTY; without even the implied
- * warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
- * See the GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this software.
- *
- * If not, see <http://www.gnu.org/licenses/>.
- */
-
-#pragma once
-
-#include "platform.h"
-
-
-#if defined(USE_QUAD_MIXER_ONLY)
-#define MAX_SUPPORTED_MOTORS 4
-#define MAX_SUPPORTED_SERVOS 1
-#else
-#ifndef MAX_SUPPORTED_MOTORS
-#define MAX_SUPPORTED_MOTORS 8
-#endif
-#define MAX_SUPPORTED_SERVOS 8
-#endif
