+++ src/main/fc/tasks.c
@@ -77,6 +77,7 @@
 #include "osd/osd.h"
 
 #include "pg/rx.h"
+#include "pg/motor.h"
 
 #include "rx/rx.h"
