+++ src/main/osd/osd_elements.c
@@ -93,8 +93,7 @@
 #include "osd/osd.h"
 #include "osd/osd_elements.h"
 
-#include "pg/pg.h"
-#include "pg/pg_ids.h"
+#include "pg/motor.h"
 
 #include "rx/rx.h"
